Makefile: give our local bin and modules preference
Kinda sucks that we have to do this, but otherwise a conflicting
`babel` executable (read: Open Babel) might get in the way and
that would be bad.

It sucks because it also breaks the "feature" where the user
could provide their own `n8-make-js` executable in the PATH
and n8-make would use that one instead of the built-in one.

However, the local `npm bin` dir is preferred so that's still
an option for the user.